pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[project]
name = "anytime"
version = "0.0.0"
description = "Peeking-safe streaming inference for A/B tests and online metrics."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.22",
  "scipy>=1.10",
  "click>=8.1",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "hypothesis>=6.88",
  "ruff>=0.5.0",
]
plot = [
  "matplotlib>=3.7",
]
demo = [
  "streamlit>=1.30",
]

[tool.hatch.build.targets.wheel]
packages = ["anytime"]

[project.scripts]
anytime = "anytime.cli.main:main"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

